Moved sus patterns into handlers folder. Updated documentation
